pyproject.toml
[project]
name = "fanbient"
version = "0.1.0"
description = "Ambient smart fan system for autonomous cooling during sleep"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sounddevice>=0.5",
    "librosa>=0.10",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "paho-mqtt>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
thermal = [
    "opencv-python>=4.9",
]
api = [
    "fastapi>=0.111",
    "uvicorn>=0.29",
]

[project.scripts]
fanbient = "fanbient.cli:app"

[project.urls]
Repository = "https://github.com/dirkleas/fanbient"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fanbient"]
